requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
multidict==6.1.0
pact-python==2.2.1
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.13.1
